feat(log): Header 컴포넌트 분리 (#6)
