+++ app/(main)/log/[id]/styles.css.ts
@@ -31,18 +31,6 @@ export const header = style({
     alignItems: 'center',
 })
 
-export const breadcrumbsWrapper = style({
-    minWidth: '220px',
-    '@media': {
-        [`screen and (max-width: 1166px)`]: {
-            display: 'none',
-        },
-        [`screen and (max-width: ${breakPoints.mobile}px)`]: {
-            display: 'none',
-        },
-    },
-})
-
 export const tableWrapper = style({
     flex: 1,
     display: 'flex',
